added parseRule util
